requirements.txt
antlr4-python3-runtime==4.13.2
cairocffi==1.7.1
CairoSVG==2.7.1
cffi==1.17.1
cssselect2==0.7.0
CTkTable==1.1
customtkinter==5.2.2
darkdetect==0.8.0
defusedxml==0.7.1
packaging==24.2
pillow==11.0.0
pycparser==2.22
python-dotenv==1.0.1
tinycss2==1.4.0
webencodings==0.5.1
